dashboard/requirements.txt
fastapi
uvicorn
jinja2
prometheus-api-client
pyyaml
pydantic